requirements.txt
python-dotenv
requests
pandas
schedule
feedparser
google-generativeai
Pillow
